move createAction into base controller
